pyproject.toml
[project]
name = "10370"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.13.0.2",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4",
    "langchain-core>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "openai>=2.6.1",
    "oracledb>=3.4.0",
    "sentence-transformers>=5.1.2",
    "snowflake-connector-python>=4.0.0",
    "snowflake-sqlalchemy>=1.1.18",
    "sqlalchemy>=1.4.54",
    "sqlglot>=27.29.0",
    "streamlit>=1.51.0",
]
